im2recipe-Pytorch_v2/requirements.txt
Cython
word2vec
numpy
scipy==1.1.0
pillow
lmdb
torchfile
scikit-learn
tqdm
nltk
simplejson
torch
torchvision
